Split sidebar Playwright flows
